Fixed an issue where apps with the library would not work on 64-bit devices.

+++ Framework Project/Colors/UIColor+ColorFinder.m
@@ -14,7 +14,7 @@
 - (UIColor *)mostSimilarColorInSet:(NSSet *)colors
 {
 	UIColor *closestColor = nil;
-	CGFloat closestDifference = FLT_MAX;
+	CGFloat closestDifference = DBL_MAX;
 	for ( UIColor *color in colors ) {
 		CGFloat difference = [UIColor differenceBetweenColor:self andColor:color];
 		if ( difference < closestDifference ) {
